requirements.txt
Django==4.0.6
djangorestframework==3.12.4
django-cors-headers==3.8.0
coreapi==2.3.3